Always cast inputs to full precision before scaling

See merge request atlas-flavor-tagging-tools/algorithms/salt!112

+++ salt/data/datasets.py
@@ -165,7 +165,7 @@ class JetDataset(Dataset):
 
     def scale_input(self, batch: dict, input_type: str):
         """Normalise jet inputs."""
-        inputs = s2u(batch[self.variables[input_type]])  # type: ignore
+        inputs = s2u(batch[self.variables[input_type]]).astype(np.float32)  # type: ignore
         if self.nan_to_num:
             inputs = np.nan_to_num(inputs)
         inputs = (inputs - self.norm[input_type]["mean"]) / self.norm[input_type]["std"]
